+++ logc_config/setting_type.c
@@ -1,0 +1,35 @@
+/* Copyright (c) 2021 CZ.NIC z.s.p.o. (http://www.nic.cz/)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+#include <logc_config.h>
+#include <stdlib.h>
+#include <string.h>
+
+bool log_check_config_setting_type(log_t log, const config_setting_t *setting, int type) {
+	int stype = config_setting_type(setting);
+	bool correct_type = stype == type;
+	if (!correct_type) {
+		char *path = config_setting_path(setting);
+		log_error(log, "The configuration has unexpected type, expected '%s' but was '%s': %s",
+			config_type_name(type), config_type_name(stype), path);
+		free(path);
+	}
+	return correct_type;
+}